Standardize text format wrapper description preprocess logic
This existing logic is the same as the GinDescriptionToggle::preprocess
function except that it does not check if we are on a gin content form
route.

This fixes an issue where the description is always visually-hidden on
non gin content form routes and thus will never have a help toggle
button rendered.